requirements.txt
streamlit>=1.28.0
vanna>=0.7.9
pandas>=2.0.0
mysql-connector-python>=8.0.0
plotly>=5.15.0
requests>=2.31.0
sqlparse>=0.4.0
langchain>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
openai>=1.0.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0